requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
distlib==0.3.8
execnet==2.1.1
filelock==3.15.4
flake8==7.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
outcome==1.3.0.post0
packaging==24.2
pep8==1.7.1
platformdirs==4.2.2
playwright==1.50.0
pluggy==1.5.0
psycopg2==2.9.9
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PySocks==1.7.1
pytest==8.3.3
pytest-cache==1.0
pytest-pep8==1.0.6
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.6
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0